changes in services

+++ global-service/src/main/java/in/cdac/university/globalService/bean/Components.java
@@ -13,6 +13,6 @@ public class Components {
 	
 	
 	private Long unumTempleCompId;
-	private Integer unumComponentOrderNo;
+	private Double unumComponentOrderNo;
 
 }
